sport_analytics-main/pyproject.toml
[project]
name = "sport-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "dotenv>=0.9.9",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "pymysql>=1.1.1",
    "werkzeug>=3.1.3",
]

[tool.ruff]
line-length = 88
target-version = "py313"
fix = true

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]
